+++ CHANGELOG.md
@@ -13,6 +13,7 @@ NOTE: This version bumps the Realm file format to version 11. It is not possible
 ### Enhancements
 * Added RemoteMongoClient functionality to `Realm.User`
 * Added `Realm.User.refreshToken`.
+* Added `Realm.Auth.EmailPassword.callResetPasswordFunction()`.
 
 ### Fixed
 * <How to hit and notice issue? what was the impact?> ([#????](https://github.com/realm/realm-js/issues/????), since v?.?.?)